# backend/docs/swagger.yaml
basePath: /
definitions:
  dehydrotationlab4_internal_app_ds.PartitionRequest:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      noise_reduction_db:
        type: number
      doctor_comment:
        $ref: '#/definitions/sql.NullString'
      fluid_deficit:
        type: number
      formed_at:
        type: string
      id:
        type: integer
      moderator:
        $ref: '#/definitions/dehydrotationlab4_internal_app_ds.User'
      moderator_id:
        type: integer
      patient_weight:
        type: number
      status:
        type: string
      user:
        $ref: '#/definitions/dehydrotationlab4_internal_app_ds.User'
      user_id:
        type: integer
    type: object
...
  dehydrotationlab4_internal_app_ds.User:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
host: localhost:8082
info:
  contact:
    email: support@partitionlab.com
    name: API Support
  description: |-
    API для управления заявками на расчет звукоизоляции перегородок

    Без авторизации доступны:
    - Регистрация и аутентификация
    - Просмотр симптомов (GET)

    С авторизацией (пользователь):
    - Управление своими заявками
    - Добавление симптомов в заявки

    С ролью модератора:
    - Все методы пользователя
    - CRUD операции с симптомами
    - Завершение/отклонение заявок
  title: Partition Lab API
  version: "4.0"
paths:
...
  /api/requests:
    get:
      consumes:
      - application/json
      description: Требуется авторизация. Обычный пользователь видит только свои заявки.
        Модератор видит все заявки.
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Дата начала (formed_at)
        in: query
        name: from
        type: string
      - description: Дата окончания (formed_at)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  type: object
                type: array
              filters:
                type: object
              total:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Получить список заявок
      tags:
      - requests
  /api/requests/{id}:
    delete:
      consumes:
      - application/json
      description: Требуется авторизация. Доступно только создателю заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  deleted:
                    type: integer
                type: object
              filters:
                type: object
              total:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Удалить заявку
      tags:
      - requests
    get:
      consumes:
      - application/json
      description: Требуется авторизация. Возвращает заявку с симптомами.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              filters:
                type: object
              total:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Получить заявку по ID
      tags:
      - requests
    put:
      consumes:
      - application/json
      description: Требуется авторизация. Доступно только владельцу черновика.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: request
        required: true
        schema:
          properties:
            doctor_comment:
              type: string
            patient_weight:
              format: float64
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.PartitionRequest'
              filters:
                type: object
              total:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Обновить заявку
      tags:
      - requests
  /api/requests/{id}/complete:
    put:
      consumes:
      - application/json
      description: Требуется авторизация модератора. Переводит заявку в статус "завершен"
        или "отклонен".
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для завершения
        in: body
        name: request
        required: true
        schema:
          properties:
            noise_reduction_db:
              format: float64
              type: number
            doctor_comment:
              type: string
            patient_weight:
              format: float64
              type: number
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.PartitionRequest'
              filters:
                type: object
              total:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Завершить или отклонить заявку
      tags:
      - requests
  /api/requests/{id}/form:
    put:
      consumes:
      - application/json
      description: Требуется авторизация. Переводит черновик в статус "сформирован".
        Доступно только владельцу.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.PartitionRequest'
              filters:
                type: object
              total:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Сформировать заявку
      tags:
      - requests
  /api/requests/cart:
    get:
      consumes:
      - application/json
      description: Требуется авторизация. Возвращает ID черновика и количество симптомов
        в нем.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  count:
                    type: integer
                  draft_id:
                    type: integer
                type: object
              filters:
                type: object
              total:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Получить корзину
      tags:
      - requests
...
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: Сессия через cookie (session_id)
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"